:hammer: cleaned up all code and added more tests

+++ app/recipe.py
@@ -17,6 +17,11 @@ def recipe(recipe_id):
          Renders recipe.html
 
     """
+    current_recipe = None
+    current_recipe_author = None
+    current_recipe_ingredients = None
+    current_recipe_methods = None
+
     try:
         current_recipe = DB_RECIPES.find_one({'_id': ObjectId(recipe_id)})
         current_recipe_author = DB_USERS.find_one({'_id': ObjectId(current_recipe['author_id'])})
@@ -25,17 +30,19 @@ def recipe(recipe_id):
     except:
         # raises a 404 error if any of these fail
         abort(404, description="Resource not found")
-
-    if session.get('USERNAME', None) is not None:
+    # Set loged in variable
+    loged_in = False
+    if session.get('USERNAME', None):
         username = session['USERNAME']
         user = DB_USERS.find_one({'username': username})
         favorites = user['favorites']
+        loged_in = True
     else:
         favorites = []
 
     return render_template('recipe.html', current_recipe=current_recipe, current_recipe_author=current_recipe_author,
                            current_recipe_ingredients=current_recipe_ingredients,
-                           current_recipe_methods=current_recipe_methods, favorites=favorites)
+                           current_recipe_methods=current_recipe_methods, favorites=favorites, loged_in=loged_in)
 
 
 ###############
@@ -44,33 +51,47 @@ def recipe(recipe_id):
 @app.route('/add_temp_recipe/<user_id>')
 def add_temp_recipe(user_id):
     """ Creates a temporary recipe record to assign to the user
-
      :return
          Redirect to edit_recipe url
-
     """
-
-    # Default recipe image
-    image_url = 'https://res.cloudinary.com/dajuujhvs/image/upload/v1591896759/gi5h7ejymbig1yybptcy.png'
-    image_url_id = 'gi5h7ejymbig1yybptcy'
-
-    temp_record = DB_RECIPES.insert_one(
-        {
-            'name': 'My Awesome Recipe',
-            'image_url': image_url,
-            'image_url_id': image_url_id,
-            'featured': 'false',
-            'current_rating': '0',
-            'total_ratings': 0,
-            'sum_ratings': 0,
-            'author_id': ObjectId(user_id),
-            'preptime_hrs': '0',
-            'preptime_min': '0',
-            'cooktime_hrs': '0',
-            'cooktime_min': '0'
-        })
-
-    return redirect(url_for('edit_recipe', user_id=user_id, recipe_id=temp_record.inserted_id))
+    # User exists in the database and Signed in
+    if session.get('USERNAME', None):
+        username = session['USERNAME']
+        # user exists in the database
+        try:
+            current_user = DB_USERS.find_one({'username': username})
+            requested_user = DB_USERS.find_one({'_id': ObjectId(user_id)})
+            # if the current logged in user is the user requested
+            if requested_user['_id'] == current_user['_id']:
+                # Default recipe image
+                image_url = 'https://res.cloudinary.com/dajuujhvs/image/upload/v1591896759/gi5h7ejymbig1yybptcy.png'
+                image_url_id = 'gi5h7ejymbig1yybptcy'
+                # create empty temp record
+                temp_record = DB_RECIPES.insert_one(
+                    {
+                        'name': 'My Awesome Recipe',
+                        'image_url': image_url,
+                        'image_url_id': image_url_id,
+                        'featured': 'false',
+                        'current_rating': '0',
+                        'total_ratings': 0,
+                        'sum_ratings': 0,
+                        'author_id': ObjectId(user_id),
+                        'preptime_hrs': '0',
+                        'preptime_min': '0',
+                        'cooktime_hrs': '0',
+                        'cooktime_min': '0'
+                    })
+                return redirect(url_for('edit_recipe', user_id=user_id, recipe_id=temp_record.inserted_id))
+            else:
+                # raises a 403 error
+                return abort(403, description="Forbidden")
+        except:
+            # raises a 404
+            return abort(404, description="Resource not found")
+    else:
+        # User not signed in
+        return redirect(url_for('sign_in'))
 
 
 ####################
@@ -79,17 +100,31 @@ def add_temp_recipe(user_id):
 @app.route('/edit_recipe/<user_id>/<recipe_id>')
 def edit_recipe(user_id, recipe_id):
     """ Renders the edit-recipe template passing in the temporary recipe id
-
      :return
          returns the edit-recipe.html form
-
     """
-    current_recipe = DB_RECIPES.find_one({'_id': ObjectId(recipe_id)})
-    current_ingredients = DB_INGREDIENTS.find({'recipe_id': current_recipe['_id']})
-    current_methods = DB_METHODS.find({'recipe_id': current_recipe['_id']})
+    # User not Signed in
+    if not session.get('USERNAME', None):
+        return redirect(url_for('sign_in'))
 
-    return render_template('edit-recipe.html', user_id=user_id, recipe_id=recipe_id, current_recipe=current_recipe,
-                           current_ingredients=current_ingredients, current_methods=current_methods)
+    username = session['USERNAME']
+    # user exists in the database
+    try:
+        current_user = DB_USERS.find_one({'username': username})
+        current_recipe = DB_RECIPES.find_one({'_id': ObjectId(recipe_id)})
+        current_ingredients = DB_INGREDIENTS.find({'recipe_id': current_recipe['_id']})
+        current_methods = DB_METHODS.find({'recipe_id': current_recipe['_id']})
+        # if the current logged in user is recipe author
+        if current_recipe['author_id'] == current_user['_id']:
+            return render_template('edit-recipe.html', user_id=user_id, recipe_id=recipe_id,
+                                   current_recipe=current_recipe,
+                                   current_ingredients=current_ingredients, current_methods=current_methods)
+        else:
+            # raises a 403 error
+            return abort(403, description="Forbidden")
+    except:
+        # raises a 404 error if any of these fail
+        return abort(404, description="Resource not found")
 
 
 #################
@@ -103,25 +138,33 @@ def update_recipe(user_id, recipe_id):
         returns the edit-recipe.html form overview tab
 
     """
-    # get the current user's record
-    current_user = DB_USERS.find_one({'_id': ObjectId(user_id)})
+    # User not Signed in
+    if not session.get('USERNAME', None):
+        return redirect(url_for('sign_in'))
 
-    # update the recipe record
-    DB_RECIPES.update_one({'_id': ObjectId(recipe_id)}, {"$set": {
-        'name': request.form.get('recipe-name'),
-        'description': request.form.get('recipe-description'),
-        'notes': request.form.get('recipe-notes'),
-        'preptime_hrs': request.form.get('prep-hours'),
-        'preptime_min': request.form.get('prep-minutes'),
-        'cooktime_hrs': request.form.get('cook-hours'),
-        'cooktime_min': request.form.get('cook-minutes'),
-        'serves': request.form.get('serves'),
-        'author': current_user['username'],
-        'author_id': ObjectId(user_id),
-        'date_updated': TODAY_STR
-    }}, upsert=True)
-    # return to recipe overview page
-    return redirect(url_for('edit_recipe', _anchor='overview', user_id=user_id, recipe_id=recipe_id))
+    try:
+        # get the current user's record
+        current_user = DB_USERS.find_one({'_id': ObjectId(user_id)})
+
+        # update the recipe record
+        DB_RECIPES.update_one({'_id': ObjectId(recipe_id)}, {"$set": {
+            'name': request.form.get('recipe-name'),
+            'description': request.form.get('recipe-description'),
+            'notes': request.form.get('recipe-notes'),
+            'preptime_hrs': request.form.get('prep-hours'),
+            'preptime_min': request.form.get('prep-minutes'),
+            'cooktime_hrs': request.form.get('cook-hours'),
+            'cooktime_min': request.form.get('cook-minutes'),
+            'serves': request.form.get('serves'),
+            'author': current_user['username'],
+            'author_id': ObjectId(user_id),
+            'date_updated': TODAY_STR
+        }}, upsert=True)
+        # return to recipe overview page
+        return redirect(url_for('edit_recipe', _anchor='overview', user_id=user_id, recipe_id=recipe_id))
+    except:
+        # raises a 404 error if any of these fail
+        return abort(404, description="Resource not found")
 
 
 #######################
@@ -135,6 +178,9 @@ def update_recipe_image(user_id, recipe_id):
         returns the edit-recipe.html form image tab with updated image
 
     """
+    # User not Signed in
+    if not session.get('USERNAME', None):
+        return redirect(url_for('sign_in'))
 
     # get the file from the form
     file_to_upload = request.files.get('file')
@@ -169,11 +215,25 @@ def del_recipe(recipe_id):
        Redirects to the user's profile url
 
    """
-    DB_RECIPES.remove({'_id': ObjectId(recipe_id)})
-    DB_INGREDIENTS.delete_many({'recipe_id': ObjectId(recipe_id)})
-    DB_METHODS.delete_many({'recipe_id': ObjectId(recipe_id)})
+    # User not Signed in
+    if not session.get('USERNAME', None):
+        return redirect(url_for('sign_in'))
+    # Get username
+    username = session['USERNAME']
+    try:
+        current_user = DB_USERS.find_one({'username': username})
+        current_recipe = DB_RECIPES.find_one({'_id': ObjectId(recipe_id)})
+        # if the current logged in user is recipe author
+        if current_recipe['author_id'] == current_user['_id']:
+            DB_RECIPES.remove({'_id': ObjectId(recipe_id)})
+            DB_INGREDIENTS.delete_many({'recipe_id': ObjectId(recipe_id)})
+            DB_METHODS.delete_many({'recipe_id': ObjectId(recipe_id)})
 
-    return redirect(url_for('profile'))
+            return redirect(url_for('profile'))
+
+    except:
+        # raises a 404 error if any of these fail
+        return abort(404, description="Resource not found")
 
 
 ##################
@@ -258,17 +318,21 @@ def rate_recipe(recipe_id):
     :return
         Redirect to recipe page
     """
-    # calculate new rating
-    current_recipe = DB_RECIPES.find_one({'_id': ObjectId(recipe_id)})
-    calculated_rating_total = int(current_recipe['total_ratings']) + 1
-    calculated_sum = int(current_recipe['sum_ratings']) + int(request.form.get('rating'))
-    # rounded average for simplicity
-    calculated_avg = round(calculated_sum / calculated_rating_total)
-    # update record
-    DB_RECIPES.update_one({'_id': ObjectId(recipe_id)}, {"$set": {
-        'total_ratings': calculated_rating_total,
-        'sum_ratings': calculated_sum,
-        'current_rating': calculated_avg
-    }}, upsert=True)
+    try:
+        # calculate new rating
+        current_recipe = DB_RECIPES.find_one({'_id': ObjectId(recipe_id)})
+        calculated_rating_total = int(current_recipe['total_ratings']) + 1
+        calculated_sum = int(current_recipe['sum_ratings']) + int(request.form.get('rating'))
+        # rounded average for simplicity
+        calculated_avg = round(calculated_sum / calculated_rating_total)
+        # update record
+        DB_RECIPES.update_one({'_id': ObjectId(recipe_id)}, {"$set": {
+            'total_ratings': calculated_rating_total,
+            'sum_ratings': calculated_sum,
+            'current_rating': calculated_avg
+        }}, upsert=True)
 
+    except:
+        # raises a 404 error if any of these fail
+        return abort(404, description="Resource not found")
     return redirect(url_for('recipe', recipe_id=recipe_id))
